+++ seaice3p/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.15.0"
+__version__ = "0.16.0"
 
 # Exported functions and classes
 from .params import *